update Navbar
getItem('token') 추가
