+++ src/components/Navbar/Navbar.scss
@@ -1,86 +1,7 @@
 @import '../../styles/variables.scss';
 
 .navbar {
+  width: 85.75rem;
   height: 100px;
-
-  .firstNavbar {
-    background: #fafafa;
-    height: 2.125rem;
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-
-    .firstNav {
-      display: flex;
-
-      div {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        position: relative;
-        cursor: pointer;
-      }
-
-      div::before {
-        content: '';
-        position: absolute;
-        bottom: 0;
-        background-color: red;
-        height: 2px;
-        width: 0;
-        transition: 0.7s;
-      }
-
-      div:hover::before {
-        width: 60%;
-      }
-
-      p {
-        padding: 0 1.5rem 0 0.125rem;
-        font-size: 0.875rem;
-      }
-    }
-  }
-
-  .secondNavbar {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: sticky;
-    top: 0;
-    width: 100%;
-    padding: 0 1.563rem;
-
-    .luluLogoSide {
-      display: flex;
-      justify-content: flex-start;
-      align-items: center;
-      flex-grow: 0.05;
-
-      .luluLogo {
-        width: 2.125rem;
-        cursor: pointer;
-      }
-    }
-    .secondNavbarCategory {
-      display: flex;
-      justify-content: flex-start;
-      align-items: center;
-      flex-grow: 0.6;
-    }
-
-    .secondNavbarRightSide {
-      display: flex;
-      justify-content: flex-end;
-      align-items: center;
-      padding: 1.125rem 0;
-      flex-grow: 0.3;
-
-      .shoppingBag {
-        width: 1.5rem;
-        margin-left: 0.5rem;
-        cursor: pointer;
-      }
-    }
-  }
+  margin: 0 auto;
 }
