+++ src/components/Navbar/NavbarSearchBox/NavbarSearchBox.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .navbarSearchBox {
   position: relative;
